refactor: hook 분리 #92
