+++ frontend/src/apis/room.ts
@@ -25,7 +25,7 @@ export const makeRoom = async (nickname: string): Promise<RoomAndMember> => {
 };
 
 // 방 참여하기
-export const enterRoom = async (roomId: number, nickname: string): Promise<RoomInfo> => {
+export const enterRoom = async (roomId: number, nickname: string): Promise<RoomAndMember> => {
   const res = await fetcher.post({
     url: API_URL.enterRoom(roomId),
     headers: {
@@ -45,14 +45,14 @@ export const enterRoom = async (roomId: number, nickname: string): Promise<RoomI
   return data;
 };
 
-// 방 전체 멤버 조회
-export const getRoomMembers = async (roomId: number): Promise<RoomMembers> => {
+// 방 정보 조회
+export const getRoomInfo = async (roomId: number): Promise<RoomInfo> => {
   const res = await fetcher.get({
-    url: API_URL.roomMembers(roomId),
+    url: API_URL.getRoomInfo(roomId),
   });
 
   if (!res.ok) {
-    throw new Error('방 멤버 가져오기에 실패하였습니다');
+    throw new Error('방 정보 가져오기에 실패하였습니다');
   }
 
   const data = await res.json();
