+++ frontend/src/pages/ReadyPage/ReadyPage.tsx
@@ -1,32 +1,14 @@
-import { useQuery } from '@tanstack/react-query';
-import { useLocation, useParams } from 'react-router-dom';
-import { useRecoilValue } from 'recoil';
-
 import { readyPageLayout } from './ReadyPage.styled';
+import { useGameStart } from './useGameStart';
+import { useGetRoomInfo } from './useGetRoomInfo';
 
-import { getRoomMembers } from '@/apis/room';
 import CategoryContainer from '@/components/CategoryContainer/CategoryContainer';
 import Button from '@/components/common/Button/Button';
 import ReadyMembersContainer from '@/components/ReadyMembersContainer/ReadyMembersContainer';
-import { QUERY_KEYS } from '@/constants/queryKeys';
-import { memberInfoState } from '@/recoil/atom';
 
 const ReadyPage = () => {
-  const memberInfo = useRecoilValue(memberInfoState);
-  const { roomId } = useParams();
-
-  const { data, isLoading, isError } = useQuery({
-    queryKey: [QUERY_KEYS.roomMembers, Number(roomId)],
-    queryFn: ({ queryKey: [_, roomId] }) => getRoomMembers(roomId as number),
-    refetchInterval: 1000,
-  });
-
-  const handleGameStart = () => {
-    if (memberInfo.isMaster) {
-      // TODO: 게임 시작 API 연결 예정
-      alert('Game Start');
-    }
-  };
+  const { data, isLoading, isError } = useGetRoomInfo();
+  const { isMaster, handleGameStart } = useGameStart();
 
   return (
     <div css={readyPageLayout}>
@@ -34,7 +16,7 @@ const ReadyPage = () => {
       {isError && <div>에러 발생</div>}
       {isLoading && <div>로딩중.......</div>}
       {data && <ReadyMembersContainer members={data.members} />}
-      <Button text="시작" disabled={!memberInfo.isMaster} onClick={handleGameStart} bottom />
+      <Button text="시작" disabled={!isMaster} onClick={handleGameStart} bottom />
     </div>
   );
 };
